+++ nautilus_trader/adaptors/ccxt/data.pyx
@@ -19,7 +19,7 @@ from cpython.datetime cimport datetime
 
 from ccxt.base.errors import BaseError as CCXTError
 
-from nautilus_trader.adapters.ccxt.providers import CCXTInstrumentProvider
+from nautilus_trader.adaptors.ccxt.providers import CCXTInstrumentProvider
 
 from nautilus_trader.common.clock cimport LiveClock
 from nautilus_trader.common.logging cimport Logger